Move user args DSL testing into separate file
