+++ tests/test_command_generation.py
@@ -1,4 +1,4 @@
-from unittest import skip, expectedFailure
+from unittest import skip, expectedFailure  # noqa: F401
 
 import sublime
 from SublimeLinter.lint import (
@@ -50,164 +50,9 @@ class _BaseTestCase(DeferrableTestCase):
 
     def setUp(self):
         ...
+
     def tearDown(self):
         ...
-
-class TestArgsDSL(_BaseTestCase):
-    @p.expand([
-        # Testing arg DSL
-        # JOINER
-        # empty joiner
-        ('-a', {'a': 'foo'}, ['-a', 'foo']),
-        ('-a', {'a': ['foo']}, ['-a', 'foo']),
-        ('-a', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        ('--a', {'a': 'foo'}, ['--a', 'foo']),
-        ('--a', {'a': ['foo']}, ['--a', 'foo']),
-        ('--a', {'a': ['foo', 'bar']}, ['--a', 'foo', '--a', 'bar']),
-        ('@a', {'a': 'foo'}, ['foo']),
-        ('@a', {'a': ['foo']}, ['foo']),
-        # The following is basically what `args` does
-        ('@a', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        # deprecate! ':' can be omitted
-        ('-a:', {'a': 'foo'}, ['-a', 'foo']),
-        ('-a:', {'a': ['foo']}, ['-a', 'foo']),
-        ('-a:', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        ('--a:', {'a': 'foo'}, ['--a', 'foo']),
-        ('--a:', {'a': ['foo']}, ['--a', 'foo']),
-        ('--a:', {'a': ['foo', 'bar']}, ['--a', 'foo', '--a', 'bar']),
-        ('@a:', {'a': 'foo'}, ['foo']),
-        ('@a:', {'a': ['foo']}, ['foo']),
-        ('@a:', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        # '=' joiner
-        ('-a=', {'a': 'foo'}, ['-a=foo']),
-        ('-a=', {'a': ['foo']}, ['-a=foo']),
-        ('-a=', {'a': ['foo', 'bar']}, ['-a=foo', '-a=bar']),
-        ('--a=', {'a': 'foo'}, ['--a=foo']),
-        ('--a=', {'a': ['foo']}, ['--a=foo']),
-        ('--a=', {'a': ['foo', 'bar']}, ['--a=foo', '--a=bar']),
-        # @ and = has no effect; '=' can be omitted
-        ('@a=', {'a': 'foo'}, ['foo']),
-        ('@a=', {'a': ['foo']}, ['foo']),
-        ('@a=', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        # SEPARATOR
-        # empty joiner
-        ('-a ', {'a': 'foo'}, ['-a', 'foo']),
-        ('-a ', {'a': ['foo']}, ['-a', 'foo']),
-        ('-a ', {'a': ['foo', 'bar']}, ['-a', 'foo bar']),
-        ('--a ', {'a': 'foo'}, ['--a', 'foo']),
-        ('--a ', {'a': ['foo']}, ['--a', 'foo']),
-        ('--a ', {'a': ['foo', 'bar']}, ['--a', 'foo bar']),
-        ('@a ', {'a': 'foo'}, ['foo']),
-        ('@a ', {'a': ['foo']}, ['foo']),
-        ('@a ', {'a': ['foo', 'bar']}, ['foo bar']),
-        # ':' joiner. Can be omitted
-        ('-a: ', {'a': 'foo'}, ['-a', 'foo']),
-        ('-a: ', {'a': ['foo']}, ['-a', 'foo']),
-        ('-a: ', {'a': ['foo', 'bar']}, ['-a', 'foo bar']),
-        ('--a: ', {'a': 'foo'}, ['--a', 'foo']),
-        ('--a: ', {'a': ['foo']}, ['--a', 'foo']),
-        ('--a: ', {'a': ['foo', 'bar']}, ['--a', 'foo bar']),
-        ('@a: ', {'a': 'foo'}, ['foo']),
-        ('@a: ', {'a': ['foo']}, ['foo']),
-        ('@a: ', {'a': ['foo', 'bar']}, ['foo bar']),
-        ('@a,', {'a': ['foo', 'bar']}, ['foo,bar']),
-        # EXCEPT if we want ':' or '=' as the separator
-        ('-a::', {'a': ['foo', 'bar']}, ['-a', 'foo:bar']),
-        ('-a:=', {'a': ['foo', 'bar']}, ['-a', 'foo=bar']),
-        # or any char allowed in name
-        ('-a:-', {'a': ['foo', 'bar']}, ['-a', 'foo-bar']),
-        ('-a:_', {'a': ['foo', 'bar']}, ['-a', 'foo_bar']),
-        ('-a:a', {'a': ['foo', 'bar']}, ['-a', 'fooabar']),
-        #
-        # '=' joiner
-        ('-a= ', {'a': 'foo'}, ['-a=foo']),
-        ('-a= ', {'a': ['foo']}, ['-a=foo']),
-        ('-a= ', {'a': ['foo', 'bar']}, ['-a=foo bar']),
-        ('--a= ', {'a': 'foo'}, ['--a=foo']),
-        ('--a= ', {'a': ['foo']}, ['--a=foo']),
-        ('--a= ', {'a': ['foo', 'bar']}, ['--a=foo bar']),
-        ('@a= ', {'a': 'foo'}, ['foo']),
-        ('@a= ', {'a': ['foo']}, ['foo']),
-        ('@a= ', {'a': ['foo', 'bar']}, ['foo bar']),
-        # multiple?
-        # Hm, doesn't do anything on normal values
-        ('-a +', {'a': 'foo'}, ['-a', 'foo']),
-        ('-a +', {'a': ['foo']}, ['-a', 'foo']),
-        # These are all the same
-        ('-a', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        ('-a +', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        ('-a,+', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        ('-a:,+', {'a': ['foo', 'bar']}, ['-a', 'foo', '-a', 'bar']),
-        #
-
-        # Hm, doesn't do anything on normal values
-        ('-a= +', {'a': 'foo'}, ['-a=foo']),
-        ('-a= +', {'a': ['foo']}, ['-a=foo']),
-        # Ignores the joiner on multiple values:
-        ('-a= +', {'a': ['foo', 'bar']}, ['-a=foo', '-a=bar']),
-        ('-a=,+', {'a': ['foo', 'bar']}, ['-a=foo', '-a=bar']),
-        # is actually the same as just
-        ('-a=', {'a': ['foo', 'bar']}, ['-a=foo', '-a=bar']),
-        #
-        # These are all the same
-        ('@a', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        ('@a= +', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        ('@a: +', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-        ('@a,+', {'a': ['foo', 'bar']}, ['foo', 'bar']),
-
-        #
-        # True value acts as a switch
-        ('-a', {'a': True}, ['-a']),
-        ('--a', {'a': True}, ['--a']),
-        ('@a', {'a': True}, ['@a']),
-        # joiner doesn't matter
-        ('-a=', {'a': True}, ['-a']),
-        ('--a=', {'a': True}, ['--a']),
-
-        #
-        # 0 (Zero) is not falsy
-        ('-a', {'a': 0}, ['-a', '0']),
-        ('--a', {'a': 0}, ['--a', '0']),
-        ('@a', {'a': 0}, ['0']),
-
-        #
-        #
-    ])
-    def test_truthy_values(self, arg, settings, result):
-        class FakeLinterArgDSL(Linter):
-            defaults = {
-                'selector': None,
-                arg: None
-            }
-            cmd = 'fake_linter_1'
-
-        linter = FakeLinterArgDSL(self.view, settings)
-        cmd = ['fake_linter_1'] + result
-        with expect(linter)._communicate(cmd, ...):
-            linter.lint(INPUT, VIEW_UNCHANGED)
-
-    @p.expand([
-        ('-a', {'a': None}),
-        ('-a', {'a': False}),
-        ('-a', {'a': ''}),
-        ('--a', {'a': None}),
-        ('--a', {'a': False}),
-        ('--a', {'a': ''}),
-        ('@a', {'a': None}),
-        ('@a', {'a': False}),
-        ('@a', {'a': ''}),
-    ])
-    def test_falsy_values(self, arg, settings):
-        class FakeLinterArgDSL(Linter):
-            defaults = {
-                'selector': None,
-                arg: None
-            }
-            cmd = 'fake_linter_1'
-
-        linter = FakeLinterArgDSL(self.view, settings)
-        with expect(linter)._communicate(['fake_linter_1'], ...):
-            linter.lint(INPUT, VIEW_UNCHANGED)
 
 
 class TestArgsSetting(_BaseTestCase):
@@ -308,6 +153,18 @@ class TestExecutableSetting(_BaseTestCase):
 # - selects first folder if no filename
 # - or dirname of file
 #
+# 'context_sensitive_executable_path' contract
+# returns
+# - True, path  --> take path
+# - True, None  --> abort linting
+# - False, None --> use can_exec and which (SL defaults)
+#
+# 'cmd'
+# - can be string
+# - can be tuple/list
+# - can be a callable which
+#   - returns a string
+#   - returns a tuple/list
 #
 
 # PythonLinter
